pyproject.toml
[project]
name = "lisp-sql-tests"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["psycopg2-binary"]

[dependency-groups]
dev = ["pytest"]
